requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
joblib==1.4.2
pandas==2.2.2
scikit-learn==1.6.0
lightgbm
shap==0.46.0
google-genai==0.3.0
python-dotenv==1.0.1
numpy>=2.0.0

google-cloud-firestore>=2.19.0
httpx>=0.27.0
openai>=1.50.0